# Enables the incremental reader to iterate annotations lazily.