*Description of changes:*
This PR includes two optimizations:
1. (Minor) Uses `IntList` introduced in #404 for storing annotation SIDs without upgrading to `Integer`.
2. Lazily parsing and resolving annotation SIDs. Previously, annotation SIDs were eagerly parsed (in `IonReaderLookaheadBuffer`) at the top-level, and parsed on-demand but in their entirety below the top-level (in `IonReaderBinaryIncremental.getAnnotationSids`). Now, as long as annotation iterator reuse is enabled, annotations at any level are both parsed and resolved lazily (when accessed via the iterator provided by `IonReader.iterateTypeAnnotations`). This enables performance improvements in cases where the user does not access some or all annotations on a value, as is the case when skip-scanning.

*Performance:*

I used ion-java-benchmark-cli to verify that there is no observed impact to performance for streams that don't make heavy use of annotations.

I also used ion-java-benchmark-cli to do a comparison targeted to this change. The test data is a stream of top-level floats (chosen to minimize the value parsing time), each with a variable number of annotations.

The first trial involved iterating all annotations on every value.

Before:

```
Benchmark                                               (input)                                              (options)  Mode  Cnt          Score       Error   Units
Bench.run                                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        632.625 ±     6.560   ms/op
Bench.run:Heap usage                        annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4         28.486 ±     0.305      MB
Bench.run:Serialized size                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        100.350                  MB
Bench.run:·gc.alloc.rate                    annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        188.526 ±     1.849  MB/sec
Bench.run:·gc.alloc.rate.norm               annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  131332260.000 ±     7.462    B/op
Bench.run:·gc.churn.PS_Eden_Space           annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        186.506 ±     1.829  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  129925120.000 ±     0.001    B/op
Bench.run:·gc.churn.PS_Survivor_Space       annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4          0.314 ±     0.066  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4     218647.250 ± 45039.772    B/op
Bench.run:·gc.count                         annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        260.000              counts
Bench.run:·gc.time                          annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4         96.000                  ms
```

After:

```
Benchmark                                  (input)                                              (options)  Mode  Cnt      Score    Error   Units
Bench.run                      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4    564.878 ±  4.420   ms/op
Bench.run:Heap usage           annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4      5.666 ±  4.339      MB
Bench.run:Serialized size      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4    100.350               MB
Bench.run:·gc.alloc.rate       annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4      0.135 ±  0.001  MB/sec
Bench.run:·gc.alloc.rate.norm  annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  84128.889 ± 11.488    B/op
Bench.run:·gc.count            annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        ≈ 0           counts
```

The second trial involved retrieving only the first annotation on each value.

Before:

```
Benchmark                                               (input)                                              (options)  Mode  Cnt          Score         Error   Units
Bench.run                                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        595.577 ±      24.995   ms/op
Bench.run:Heap usage                        annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4         19.235 ±      94.244      MB
Bench.run:Serialized size                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        100.350                    MB
Bench.run:·gc.alloc.rate                    annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        200.281 ±       8.033  MB/sec
Bench.run:·gc.alloc.rate.norm               annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  131332258.353 ±       7.023    B/op
Bench.run:·gc.churn.PS_Eden_Space           annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        199.079 ±       3.125  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  130547711.882 ± 5544439.327    B/op
Bench.run:·gc.churn.PS_Survivor_Space       annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4          0.303 ±       0.077  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4     199042.118 ±   56985.986    B/op
Bench.run:·gc.count                         annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        258.000                counts
Bench.run:·gc.time                          annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        118.000                    ms
```

After:

```
Benchmark                                  (input)                                              (options)  Mode  Cnt      Score   Error   Units
Bench.run                      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4    497.792 ± 8.578   ms/op
Bench.run:Heap usage           annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4      5.330 ± 0.028      MB
Bench.run:Serialized size      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4    100.350              MB
Bench.run:·gc.alloc.rate       annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4      0.154 ± 0.003  MB/sec
Bench.run:·gc.alloc.rate.norm  annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4  84125.333 ± 5.685    B/op
Bench.run:·gc.count            annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:INCREMENTAL}  avgt    4        ≈ 0          counts
```

For reference, here are the results for the incremental reader (which had no performance difference when reading a single annotation vs all annotations because annotations are parsed and resolved eagerly):

```
Benchmark                                               (input)                                                  (options)  Mode  Cnt          Score         Error   Units
Bench.run                                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        832.241 ±      19.998   ms/op
Bench.run:Heap usage                        annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4         27.072 ±      44.461      MB
Bench.run:Serialized size                   annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        100.350                    MB
Bench.run:·gc.alloc.rate                    annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        199.936 ±       7.366  MB/sec
Bench.run:·gc.alloc.rate.norm               annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4  182990436.615 ±      26.216    B/op
Bench.run:·gc.churn.PS_Eden_Space           annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        198.118 ±       8.860  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4  181326347.987 ± 3377440.465    B/op
Bench.run:·gc.churn.PS_Survivor_Space       annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4          0.232 ±       0.043  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4     212396.051 ±   39487.424    B/op
Bench.run:·gc.count                         annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        282.000                counts
Bench.run:·gc.time                          annotatedFloats.10n  read::{f:ION_BINARY,t:FILE,a:STREAMING,R:NON_INCREMENTAL}  avgt    4        113.000                    ms
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.